Makes username/email case insensitive (#51)
This adds a migration to downcase all username/emails. It also adds a
before_save hook to downcase email/usernames. Last but not least, it
adds an extra saftey measure of `lower(username) OR lower(email)` to the
SQL query and to the username/email sent for login. Only way this would
ever conflict is if the migration is not installed and two users have
the same email or username.